requirements.txt
requests
flask
python-telegram-bot
apscheduler
bs4
sqlalchemy
psycopg2